+++ web/apps/photos/src/pages/gallery.tsx
@@ -13,6 +13,7 @@ import {
     CollectionSelector,
     type CollectionSelectorAttributes,
 } from "@/new/photos/components/CollectionSelector";
+import { PlanSelector } from "@/new/photos/components/PlanSelector";
 import {
     SearchBar,
     type SearchBarProps,
@@ -104,7 +105,6 @@ import Sidebar from "components/Sidebar";
 import { type UploadTypeSelectorIntent } from "components/Upload/UploadTypeSelector";
 import Uploader from "components/Upload/Uploader";
 import { UploadSelectorInputs } from "components/UploadSelectorInputs";
-import { PlanSelector } from "components/pages/gallery/PlanSelector";
 import SelectedFileOptions from "components/pages/gallery/SelectedFileOptions";
 import { t } from "i18next";
 import { useRouter, type NextRouter } from "next/router";
